RUST REWRITE
As the project grew from a lightweight FFmpeg CLI wrapper,
architectural issues started to appear which called for a more
vertically integrated design. And since Go is unsuitable for FFI
and assembly-level video processing, I decided to rewrite it in Rust.

Most of the things we used from FFmpeg besides the h264 decoder have
been ported to Rust.